improve: 구조 리펙토링(문자열 분리 class 생성)
